% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colors.R
\name{check_colors}
\alias{check_colors}
\title{Check Manual Colors}
\usage{
check_colors(colors)
}
\arguments{
\item{colors}{Colors to check.}
}
\value{
Hex code of input colors.
}
\description{
Check Manual Colors
}
\examples{
if (requireNamespace("unikn", quietly = TRUE))
  check_colors(c("red", "blue"))
}
\seealso{
Other rwif-plot: 
\code{\link{plot_group_dots}()},
\code{\link{plot_parallel_coordinate}()}
}
\concept{rwif-plot}
