% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_dots.R
\name{plot_group_dots}
\alias{plot_group_dots}
\title{Draw Grouped Dots}
\source{
Adapted from \url{https://mp.weixin.qq.com/s/rcnkEY_FR9BStkEEC3ms_Q}
}
\usage{
plot_group_dots(
  data,
  step = 0.25,
  small_step = 0.01,
  title = "",
  dot_alpha = 0.5,
  legend_title = "",
  legend_ncol = 1,
  pal = hcl.colors(nrow(data), "Zissou")
)
}
\arguments{
\item{data}{A numeric \code{data.frame}.}

\item{small_step}{A small step to discrete the points with equal value
for better visualization.}

\item{title, legend_title, legend_ncol, pal, step, dot_alpha}{Plot settings.}
}
\value{
Nothing.
}
\description{
Draw Grouped Dots
}
\examples{
plot_group_dots(
  VADeaths,
  title = "deaths per 1000 in 1940 Virginia stratified by age, gender, and location",
  legend_title = "age bins")
}
\seealso{
Other rwif-plot: 
\code{\link{plot_parallel_coordinate}()}
}
\concept{rwif-plot}
