% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwif-package.R
\docType{package}
\name{rwif-package}
\alias{rwif}
\alias{rwif-package}
\title{rwif: An R Companion for Data Science and Omics Analysis}
\description{
Provides personal common utils for data science and omics analysis, enabling simplicity, reproducibility and productivity.
}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
